Brian's changes
